# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Finance Copilot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FinanceAccountType = typeof FinanceAccountType[keyof typeof FinanceAccountType];


export const FinanceAccountType = {
  checking: 'checking',
  savings: 'savings',
  credit: 'credit',
  investment: 'investment',
} as const;

export interface FinanceAccount {
  id: string;
  name: string;
  institution: string;
  type: FinanceAccountType;
  mask: string;
  balance: number;
  balanceChange: number;
  balanceChangeLabel: string;
  accent: string;
}

export type FinanceAccountInputType = typeof FinanceAccountInputType[keyof typeof FinanceAccountInputType];


export const FinanceAccountInputType = {
  checking: 'checking',
  savings: 'savings',
  credit: 'credit',
  investment: 'investment',
} as const;

export interface FinanceAccountInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  institution: string;
  type: FinanceAccountInputType;
  /**
     * @minLength 4
     * @maxLength 4
     * @pattern ^[0-9]{4}$
     */
  mask: string;
  balance: number;
  balanceChange?: number;
  balanceChangeLabel?: string;
  accent?: string;
}

export type FinanceTransactionType = typeof FinanceTransactionType[keyof typeof FinanceTransactionType];


export const FinanceTransactionType = {
  income: 'income',
  expense: 'expense',
} as const;

export type FinanceTransactionStatus = typeof FinanceTransactionStatus[keyof typeof FinanceTransactionStatus];


export const FinanceTransactionStatus = {
  posted: 'posted',
  pending: 'pending',
} as const;

export interface FinanceTransaction {
  id: string;
  merchant: string;
  category: string;
  date: string;
  amount: number;
  type: FinanceTransactionType;
  account: string;
  note: string;
  status: FinanceTransactionStatus;
}

export type FinanceTransactionInputType = typeof FinanceTransactionInputType[keyof typeof FinanceTransactionInputType];


export const FinanceTransactionInputType = {
  income: 'income',
  expense: 'expense',
} as const;

export interface FinanceTransactionInput {
  /** @minLength 1 */
  merchant: string;
  /** @minLength 1 */
  category: string;
  date: string;
  amount: number;
  type: FinanceTransactionInputType;
  account: string;
  note?: string;
}

export type FinanceBudgetStatus = typeof FinanceBudgetStatus[keyof typeof FinanceBudgetStatus];


export const FinanceBudgetStatus = {
  on_track: 'on_track',
  watch: 'watch',
  over: 'over',
} as const;

export interface FinanceBudget {
  id: string;
  category: string;
  spent: number;
  limit: number;
  color: string;
  status: FinanceBudgetStatus;
}

export interface FinanceBudgetInput {
  /** @minLength 1 */
  category: string;
  /** @exclusiveMinimum 0 */
  limit: number;
  color?: string;
}

export interface FinanceTrendPoint {
  label: string;
  income: number;
  spending: number;
}

export type FinanceDashboardSpendingByCategory = {[key: string]: number};

export interface FinanceDashboard {
  totalBalance: number;
  balanceChange: number;
  balanceChangePercent: number;
  monthlyIncome: number;
  monthlySpending: number;
  savingsRate: number;
  budgetUsed: number;
  budgetLimit: number;
  insight: string;
  insightDetail: string;
  spendingByCategory: FinanceDashboardSpendingByCategory;
  trend: FinanceTrendPoint[];
  recentTransactions: FinanceTransaction[];
}

export type AssistantChatMessageRole = typeof AssistantChatMessageRole[keyof typeof AssistantChatMessageRole];


export const AssistantChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface AssistantChatMessage {
  role: AssistantChatMessageRole;
  content: string;
}

export interface AssistantChatInput {
  /** @minLength 1 */
  message: string;
  history?: AssistantChatMessage[];
}

export interface AssistantChatResponse {
  message: string;
  model: string;
  sources: string[];
}

export type ListFinanceTransactionsParams = {
search?: string;
category?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};