# lib/api-zod/src/generated/types/financeBudgetInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Finance Copilot API
 * OpenAPI spec version: 0.1.0
 */

export interface FinanceBudgetInput {
  /** @minLength 1 */
  category: string;
  /** @exclusiveMinimum 0 */
  limit: number;
  color?: string;
}